+++ source/extensions/isaacsim.robot.policy.examples/isaacsim/robot/policy/examples/interactive/franka/franka_example_extension.py
@@ -16,49 +16,37 @@
 import os
 
 import omni.ext
+from isaacsim.examples.base.base_sample_extension_experimental import BaseSampleUITemplate
 from isaacsim.examples.browser import get_instance as get_browser_instance
-from isaacsim.examples.interactive.base_sample import BaseSampleUITemplate
-from isaacsim.examples.interactive.quadruped import QuadrupedExample
+from isaacsim.robot.policy.examples.interactive.franka import FrankaExample
 
 
-class QuadrupedExampleExtension(omni.ext.IExt):
+class FrankaExampleExtension(omni.ext.IExt):
     def on_startup(self, ext_id: str):
-        self.example_name = "Quadruped"
+        self.example_name = "Franka"
         self.category = "Policy"
 
-        overview = "This Example shows an Boston Dynamics Spot running a flat terrain policy trained in Isaac Lab"
-        overview += "\n\tKeybord Input:"
-        overview += "\n\t\tup arrow / numpad 8: Move Forward"
-        overview += "\n\t\tdown arrow/ numpad 2: Move Reverse"
-        overview += "\n\t\tleft arrow/ numpad 4: Move Left"
-        overview += "\n\t\tright arrow / numpad 6: Move Right"
-        overview += "\n\t\tN / numpad 7: Spin Counterclockwise"
-        overview += "\n\t\tM / numpad 9: Spin Clockwise"
-
-        overview += "\n\nPress the 'Open in IDE' button to view the source code."
+        overview = "This Example shows a Franka Panda open drawer policy trained in Isaac Lab. "
+        overview += "The Franka will attempt to open the drawer in front of it and hold it open. "
+        overview += "The scene will reset every 10s (sim time) and the Franka will try again."
 
         ui_kwargs = {
             "ext_id": ext_id,
             "file_path": os.path.abspath(__file__),
-            "title": "Quadruped: Boston Dynamics Spot",
+            "title": "Manipulator: Franka",
             "doc_link": "https://docs.isaacsim.omniverse.nvidia.com/latest/isaac_lab_tutorials/tutorial_policy_deployment.html",
             "overview": overview,
-            "sample": QuadrupedExample(),
+            "sample": FrankaExample(),
         }
 
         ui_handle = BaseSampleUITemplate(**ui_kwargs)
 
-        # register the example with examples browser
+        # Register the example with examples browser
         get_browser_instance().register_example(
             name=self.example_name,
-            execute_entrypoint=ui_handle.build_window,
             ui_hook=ui_handle.build_ui,
             category=self.category,
         )
 
-        return
-
     def on_shutdown(self):
         get_browser_instance().deregister_example(name=self.example_name, category=self.category)
-
-        return