+++ source/standalone_examples/api/isaacsim.robot.policy.examples/spot_standalone.py
@@ -20,29 +20,37 @@ simulation_app = SimulationApp({"headless": False})
 import argparse
 
 import carb
-import numpy as np
-from isaacsim.core.api import World
-from isaacsim.core.utils.prims import define_prim
+import omni.timeline
+from isaacsim.core.deprecation_manager import import_module
+from isaacsim.core.experimental.utils.stage import define_prim
+from isaacsim.core.rendering_manager import RenderingManager
+from isaacsim.core.simulation_manager import SimulationManager
+from isaacsim.core.simulation_manager.impl.isaac_events import IsaacEvents
 from isaacsim.robot.policy.examples.robots import SpotFlatTerrainPolicy
 from isaacsim.storage.native import get_assets_root_path
+
+torch = import_module("torch")
+
 
 first_step = True
 reset_needed = False
 
-parser = argparse.ArgumentParser()
+parser = argparse.ArgumentParser(description="Select simulation device.")
 parser.add_argument("--test", default=False, action="store_true", help="Run in test mode")
+parser.add_argument("--device", type=str, choices=["cpu", "cuda"], default="cpu", help="Simulation device")
+
 args, unknown = parser.parse_known_args()
+print(f"Using device: {args.device}")
 
 
 # initialize robot on first step, run robot advance
-def on_physics_step(step_size) -> None:
+def on_physics_step(step_size, context) -> None:
     global first_step
     global reset_needed
     if first_step:
         spot.initialize()
         first_step = False
     elif reset_needed:
-        my_world.reset(True)
         reset_needed = False
         first_step = True
     else:
@@ -50,7 +58,6 @@ def on_physics_step(step_size) -> None:
 
 
 # spawn world
-my_world = World(stage_units_in_meters=1.0, physics_dt=1 / 500, rendering_dt=1 / 50)
 assets_root_path = get_assets_root_path()
 if assets_root_path is None:
     carb.log_error("Could not find Isaac Sim assets folder")
@@ -60,38 +67,50 @@ prim = define_prim("/World/Ground", "Xform")
 asset_path = assets_root_path + "/Isaac/Environments/Grid/default_environment.usd"
 prim.GetReferences().AddReference(asset_path)
 
+# spawn physics scene
+define_prim("/World/PhysicsScene", "PhysicsScene")
+
+# set rendering manager
+RenderingManager.set_dt(8.0 / 200.0)
+
+# spawn simulation manager
+SimulationManager.set_physics_sim_device(args.device)
+SimulationManager.set_physics_dt(1.0 / 200.0)
+
 # spawn robot
 spot = SpotFlatTerrainPolicy(
     prim_path="/World/Spot",
-    name="Spot",
-    position=np.array([0, 0, 0.8]),
+    position=[0, 0, 0.8],
 )
-my_world.reset()
-my_world.add_physics_callback("physics_step", callback_fn=on_physics_step)
-
 # robot command
-base_command = np.zeros(3)
+base_command = torch.zeros(3, device=args.device)
+
+# register physics callback
+_physics_callback_id = SimulationManager.register_callback(on_physics_step, IsaacEvents.POST_PHYSICS_STEP)
+
+# play simulation
+omni.timeline.get_timeline_interface().play()
+simulation_app.update()
 
 i = 0
 while simulation_app.is_running():
-    my_world.step(render=True)
-    if my_world.is_stopped():
-        reset_needed = True
-    if my_world.is_playing():
+    simulation_app.update()
+    if SimulationManager.is_simulating():
         if i >= 0 and i < 80:
             # forward
-            base_command = np.array([2, 0, 0])
+            base_command = torch.tensor([2, 0, 0], device=args.device)
         elif i >= 80 and i < 130:
             # rotate
-            base_command = np.array([1, 0, 2])
+            base_command = torch.tensor([1, 0, 2], device=args.device)
         elif i >= 130 and i < 200:
             # side ways
-            base_command = np.array([0, 1, 0])
+            base_command = torch.tensor([0, 1, 0], device=args.device)
         elif i == 200:
             i = 0
             if args.test is True:
                 print("Reached: ", spot.robot.get_world_pose()[0])
                 break
         i += 1
-
+    else:
+        reset_needed = True
 simulation_app.close()